Clicking the "Last" history button should open the most recent notice, not only the first notice ever opened. In addition, when using the "push": "top" stack option, the most recent notice is at the opposite end of the array so this should be accounted for.
Note that there was a bug in the previous implementation because $.merge() returns a new instance so the captured notices_data variable refers to an array containing only the first opened notification.

Using the defaults to get the stack value because in theory different notifications could be opened on different stacks, so we have to really on the default to determine if we are pushing to the top.