Update pnotify.core.js
The mouseleave event which is setup to reset the timer doesn't catch the close click. If you click close and quickly move the cursor out the close button hides but the message is still present until the timer runs out. Add a check to see if the message is not animating to restart the timer.